+++ packages/frontend/src/components/project/NewsSection.tsx
@@ -1,8 +1,8 @@
 import React from 'react'
 
-import { OutLink } from '../../../components'
-import { LinkIcon } from '../../../components/icons'
-import { Section } from '../../../components/project/Section'
+import { LinkIcon } from '../icons'
+import { OutLink } from '../OutLink'
+import { Section } from './Section'
 
 export interface NewsSectionProps {
   news: NewsItem[]
