+++ packages/frontend/src/pages/scaling-projects/props/getNewsSection.ts
@@ -1,7 +1,7 @@
 import { Layer2 } from '@l2beat/config'
 
+import { NewsSectionProps } from '../../../components/project/NewsSection'
 import { formatDate } from '../../../utils'
-import { NewsSectionProps } from '../view/NewsSection'
 
 export function getNewsSection(project: Layer2): NewsSectionProps {
   const news = project.details.news.map((x) => ({
